[arcane,utils] Déplace la classe 'NumArrayContainer' dans son propre fichier.
