+++ arcane/src/arcane/tests/accelerator/ArcaneTestStandaloneAcceleratorMng.cc
@@ -1,14 +1,16 @@
 ﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
 //-----------------------------------------------------------------------------
-// Copyright 2000-2022 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
+// Copyright 2000-2023 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: Apache-2.0
 //-----------------------------------------------------------------------------
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-#include "arcane/launcher/ArcaneLauncher.h"
 #include "arcane/utils/NumArray.h"
+#include "arcane/utils/PlatformUtils.h"
+
+#include "arcane/launcher/ArcaneLauncher.h"
 #include "arcane/accelerator/core/IAcceleratorMng.h"
 
 #include "arcane/accelerator/NumArrayViews.h"
